+++ src/routes/UDPCreateRoute.js
@@ -42,7 +42,7 @@ const UDPCreateRoute = ({
   if (!hasPerms) return <NoPermissionMessage />;
 
   if (fetchIsPending()) {
-    return <LoadingPane onClose={handleClose} />;
+    return <LoadingPane dismissible onClose={handleClose} />;
   }
 
   return (